[refactor] 검색 키워드 빈 문자열 검사 공통 함수 분리 & 사용자 정보 수정 공통 컴포넌트 분리
