+++ src/components/setting/UpdateProfile.tsx
@@ -1,0 +1,243 @@
+import React, {useEffect, useState} from 'react';
+import styled from 'styled-components/native';
+import {
+  fontPercentage as fp,
+  widthPercentage as wp,
+  heightPercentage as hp,
+} from '~/components/common/ResponsiveSize';
+import {useNavigation} from '@react-navigation/native';
+import {RootStackNavigationProp} from '~/App';
+import BackView from '~/components/common/BackView';
+import {GoogleIcon, KakaoIcon, NaverIcon} from '~/assets/images';
+import ImageResizer from '@bam.tech/react-native-image-resizer';
+import {TouchableOpacity} from 'react-native';
+import {useUpdateUserInfo} from '~/api/queries/auth';
+import {showToast} from '~/components/common/modal/toastConfig';
+import LoadingModal from '~/components/common/modal/LoadingModal';
+import EditNickname from '~/screens/setting/updateProfile/EditNickname';
+import EditArtCategory from '~/screens/setting/updateProfile/EditArtCategory';
+import EditPicture from '~/screens/setting/updateProfile/EditPicture';
+
+type InitProfile = {
+  favoriteArt: string;
+  nickname: string;
+  profile: string | undefined;
+  email: string;
+  providerType: string;
+};
+
+type UpdateProfileMessage = {
+  errorMsg: string;
+  successMsg: string;
+};
+
+interface UpdateProfileProps {
+  title: string;
+  initProfile: InitProfile;
+  messages: UpdateProfileMessage;
+  navigateTo: string;
+}
+
+const UpdateProfile: React.FC<UpdateProfileProps> = ({
+  title,
+  initProfile,
+  messages,
+  navigateTo,
+}) => {
+  const navigation = useNavigation<RootStackNavigationProp>();
+  const [art, setArt] = useState<string>(initProfile.favoriteArt);
+  const [nicknameKeyword, setNicknameKeyword] = useState<string>(
+    initProfile.nickname,
+  );
+  const [imageUri, setImageUri] = useState<string | undefined>(
+    initProfile.profile
+      ? `data:image/png;base64,${initProfile.profile}`
+      : initProfile.profile,
+  );
+  const [createFormData, setCreateFormData] = useState<FormData | null>(null);
+  const [isVerified, setIsVerified] = useState<boolean>(false);
+  const [isLoadingOpen, setIsLoadingOpen] = useState<boolean>(false);
+  const {
+    mutate: updateUserInfo,
+    isLoading,
+    isError,
+    isSuccess,
+  } = useUpdateUserInfo(createFormData);
+
+  useEffect(() => {
+    if (createFormData !== null) {
+      updateUserInfo();
+    }
+  }, [createFormData]);
+
+  useEffect(() => {
+    if (isError) {
+      showToast(messages.errorMsg);
+    }
+    if (isLoading) {
+      setIsLoadingOpen(true);
+    }
+    if (!isLoading) {
+      setIsLoadingOpen(false);
+    }
+    if (isSuccess) {
+      setCreateFormData(null);
+      showToast(messages.successMsg);
+      if (navigateTo === 'back') {
+        navigation.goBack();
+      } else if (navigateTo === 'Main') {
+        navigation.navigate('Main');
+      }
+    }
+  }, [isError, isLoading, isSuccess]);
+
+  const onPressComplete = async () => {
+    const formData = new FormData();
+
+    formData.append('nickname', nicknameKeyword);
+    formData.append('favoriteArt', art);
+
+    const isImage = imageUri?.search('file://');
+    if (isImage && isImage !== -1) {
+      const resizedImage = await ImageResizer.createResizedImage(
+        imageUri ?? '', // path
+        300, // width
+        300, // height
+        'JPEG', // format
+        100, // quality
+        undefined, // rotation
+        // uploadFileName, // outputPath
+        undefined, // keepMeta,
+        undefined, // options => object
+      );
+      const uri = resizedImage.uri;
+      const filename = uri.split('/').pop();
+      const match = /\.(\w+)$/.exec(filename || '');
+      const type = match ? `image/${match[1]}` : `image`;
+
+      formData.append('profile', {
+        name: filename,
+        type,
+        uri: uri,
+      });
+    }
+    setCreateFormData(formData);
+  };
+
+  return (
+    <Container>
+      <BackView title={title} line={true} />
+
+      {/* body */}
+      <Contents>
+        {/* 닉네임 */}
+        <EditNickname
+          getNickname={nicknameKeyword}
+          setNickname={setNicknameKeyword}
+          isVerified={isVerified}
+          setIsVerified={setIsVerified}
+        />
+        {/* 좋아하는 전시 분야 */}
+        <EditArtCategory getValue={art} setValue={setArt} />
+        {/* 프로필 */}
+        <EditPicture imageUri={imageUri} setImageUri={setImageUri} />
+        {/* 이메일 */}
+        <ContentColumn>
+          <SectionName>이메일</SectionName>
+          <BoxView color={true}>
+            {initProfile.providerType === 'naver' ? (
+              <NaverIcon width={20} />
+            ) : initProfile.providerType === 'gmail' ||
+              initProfile.providerType === 'google' ? (
+              <GoogleIcon width={20} />
+            ) : (
+              <KakaoIcon width={20} />
+            )}
+            <EmailText>{initProfile.email}</EmailText>
+          </BoxView>
+        </ContentColumn>
+        {/* 완료 버튼 */}
+        <TouchableOpacity
+          disabled={!(nicknameKeyword !== '' && art !== '' && isVerified)}
+          onPress={onPressComplete}>
+          <CompleteButton
+            complete={nicknameKeyword !== '' && art !== '' && isVerified}>
+            {/* 중복 확인 완료도 검사 */}
+            완료
+          </CompleteButton>
+        </TouchableOpacity>
+      </Contents>
+      {isLoadingOpen && <LoadingModal message={'정보 수정 중 :)'} />}
+    </Container>
+  );
+};
+
+export default UpdateProfile;
+
+/** style */
+const Container = styled.View`
+  flex: 1;
+`;
+
+const Contents = styled.View`
+  flex: 1;
+  width: 100%;
+  height: 100%;
+  flex-direction: column;
+  background-color: #f6f6f6;
+  padding-left: ${wp(15)}px;
+  padding-right: ${wp(15)}px;
+  padding-top: ${hp(10)}px;
+  padding-bottom: ${hp(10)}px;
+  gap: ${hp(15)}px;
+`;
+
+const ContentColumn = styled.View`
+  flex-direction: column;
+  width: 100%;
+  gap: ${hp(10)}px;
+`;
+
+const SectionName = styled.Text`
+  font-size: ${fp(19)}px;
+  color: #3c4045;
+  font-family: 'omyu pretty';
+`;
+
+interface ContentProps {
+  color: boolean;
+}
+
+const BoxView = styled.View<ContentProps>`
+  border-width: 1.5px;
+  border-color: #d3d3d3;
+  border-radius: 10px;
+  flex-direction: row;
+  background-color: ${(props: ContentProps) =>
+    props.color ? '#d9d9d9' : '#f6f6f6'};
+  padding: ${hp(10)}px;
+  gap: 6.5px;
+  align-items: center;
+`;
+
+const EmailText = styled.Text`
+  font-size: ${fp(18)}px;
+  color: white;
+  font-family: 'omyu pretty';
+  text-align: center;
+`;
+
+interface CompleteButtonProps {
+  complete: boolean;
+}
+
+const CompleteButton = styled.Text<CompleteButtonProps>`
+  padding: ${hp(10)}px;
+  border-radius: 5px;
+  text-align: center;
+  background-color: ${(props: CompleteButtonProps) =>
+    props.complete ? '#ff6f61' : '#D3D3D3'};
+  color: white;
+  font-size: ${fp(17)}px;
+  font-family: 'omyu pretty';
+`;
